.history/requirements_20250121171105.txt
Flask
psycopg2
pytesseract
Pillow
